# Fix fallback to sort aggregation for grouping only hash aggregate