--- a/cpp/benchmarks/CMakeLists.txt
+++ b/cpp/benchmarks/CMakeLists.txt
@@ -173,6 +173,7 @@ ConfigureBench(FILL_BENCH filling/repeat_benchmark.cpp)
 ConfigureBench(
   GROUPBY_BENCH groupby/group_sum_benchmark.cu groupby/group_nth_benchmark.cu
   groupby/group_shift_benchmark.cu groupby/group_struct_benchmark.cu
+  groupby/group_no_requests_benchmark.cu
 )
 
 # ##################################################################################################