--- a/cpp/benchmarks/groupby/group_no_requests_benchmark.cu
+++ b/cpp/benchmarks/groupby/group_no_requests_benchmark.cu
@@ -61,7 +61,10 @@ void BM_basic_no_requests(benchmark::State& state)
   }
 }
 
-BENCHMARK_DEFINE_F(Groupby, BasicNoRequest)(::benchmark::State& state) { BM_basic_no_requests(state); }
+BENCHMARK_DEFINE_F(Groupby, BasicNoRequest)(::benchmark::State& state)
+{ 
+  BM_basic_no_requests(state); 
+}
 
 BENCHMARK_REGISTER_F(Groupby, BasicNoRequest)
   ->UseManualTime()
@@ -99,7 +102,10 @@ void BM_pre_sorted_no_requests(benchmark::State& state)
   }
 }
 
-BENCHMARK_DEFINE_F(Groupby, PreSortedNoRequests)(::benchmark::State& state) { BM_pre_sorted_no_requests(state); }
+BENCHMARK_DEFINE_F(Groupby, PreSortedNoRequests)(::benchmark::State& state)
+{ 
+  BM_pre_sorted_no_requests(state); 
+}
 
 BENCHMARK_REGISTER_F(Groupby, PreSortedNoRequests)
   ->UseManualTime()