@abellina 2021-12-12
Trying to figure out why this failure:

```
>>>> PASSED: clang format check
/bin/bash: warning: setlocale: LC_ALL: cannot change locale (C.UTF-8)
>>>> PASSED: libcudf header existence conda/recipes/libcudf/meta.yaml check
>>>> PASSED: libcudf header existence conda/recipes/libcudf/meta.yaml check
Build step 'Execute shell' marked build as failure
[Set GitHub commit status (universal)] ERROR on repos [GHRepository@6e42889e[nodeId=MDEwOlJlcG9zaXRvcnk5MDUwNjkxOA==,description=cuDF - GPU DataFrame Library ,homepage=http://rapids.ai,name=cudf,fork=false,archived=false,size=93286,milestones={},language=C++,commits={},source=<null>,parent=<null>,isTemplate=false,url=https://api.github.com/repos/rapidsai/cudf,id=90506918,nodeId=<null>,createdAt=2017-05-07T03:43:37Z,updatedAt=2021-12-12T11:11:15Z]] (sha:10afade) with context:gpuCI/cudf/check/style
```