@abellina 2021-12-12
I ran: `python3 ./cpp/scripts/run-clang-format.py -inplace` but  no diff or errors seen, so I am not sure where the error above comes from.