@ttnghia 2021-12-12
```
cmake-format.............................................................Failed
```
That's not cpp style issue. It's cmake issue, maybe you need to check CmakeList.txt to remove spaces etc.