@jrhemstad 2021-12-12
> a grouping only (no aggregation request)

I agree the logic needs to be fixed, but can you elaborate on what the use case is for a `groupby::aggregate` without an aggregation? I wasn't aware this is something that was even supported.