@abellina 2021-12-13
> can you elaborate on what the use case is for a groupby::aggregate without an aggregation? I wasn't aware this is something that was even supported.

This type of call is done when we are trying to compute a `DISTINCT` aggregation. 

Spark implements the first aggregation  for the distinct (the one finding the unique rows) by performing a group-by where the "keys" are both the requested set of keys (if the user specified `GROUP BY`) and the columns being aggregated (`select count(distinct foo)` would add `foo` to the keys). The product of this first group-by without aggregate functions are distinct rows we can then use to perform an aggregate (like count, sum, etc).

So for an example such as: `select distinct count(my_col) from my_table`. This doesn't have a `GROUP BY` clause (that's also an option), but it boils down to a group-by aggregate in cuDF terms:

1. Perform a group-by with `my_col` as the key. Do not request a count aggregation yet as we would be counting non-distinct. This is the call that sees no aggregation requests to cuDF. 
2. Partition + shuffle so we can distribute this and reduce.
3. Perform an aggregate where we `count` these unique rows. This is a reduction for this example, but in the case where the user specified a `GROUP BY` it would be another group-by aggregate, since we'd care about the grouping key.