@jrhemstad 2021-12-13
> This type of call is done when we are trying to compute a `DISTINCT` aggregation.

Wouldn't you just use `drop_duplicates` to get the distinct keys? 

https://github.com/rapidsai/cudf/blob/d23bcb475273d6874bf797b4e77a90d0c478f9ef/cpp/include/cudf/stream_compaction.hpp#L238

There's a lot of extra machinery going on in a groupby that doing this operation through a groupby w/ no aggregation would be possibly inefficient or this functionality could be removed, etc.