@jlowe 2021-12-13
> Wouldn't you just use drop_duplicates to get the distinct keys?

drop_duplicates performs a stable sort which seems much more expensive in practice than the hash-based approach. The large perf regression caused by accidentally falling back to a sort-based aggregation implies this as well.